landing page & page.tsx changed to dashboard
