+++ client/src/app/signup/page.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { SignUp } from '@clerk/nextjs';
+
+export default function SignUpPage() {
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-white">
+      <SignUp path="/sign-up" routing="path" redirectUrl="/dashboard" />
+    </div>
+  );
+}
