+++ client/src/app/singin/page.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { SignIn } from '@clerk/nextjs';
+
+export default function SignInPage() {
+  return (
+    <div className="min-h-screen flex items-center justify-center bg-white">
+      <SignIn path="/sign-in" routing="path" redirectUrl="/dashboard" />
+    </div>
+  );
+}
